Rename mock test to testing
