+++ crates/pallet-bioauth/src/mock/mod.rs
@@ -1,4 +1,4 @@
 #[cfg(feature = "runtime-benchmarks")]
 pub(crate) mod benchmarking;
 
-pub(crate) mod test;
+pub(crate) mod testing;
